pyproject.toml
[project]
name = "monte-carlo-mag"
version = "0.1.0"
description = "Kinetic Monte Carlo simulation of magnetic nanoparticles with antibody-antigen binding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

visualization = [
    "numpy>=1.24",
    "matplotlib>=3.7",
]
